requirements.txt
adal==1.2.1
asn1crypto==0.24.0
cachetools==3.1.0
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
cryptography==2.6.1
google-auth==1.6.3
idna==2.8
kubernetes==8.0.1
oauthlib==3.0.1
pyasn1==0.4.5
pyasn1-modules==0.2.4
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.0
PyYAML==5.1
requests==2.21.0
requests-oauthlib==1.2.0
rsa==4.0
six==1.12.0
tabulate==0.8.3
urllib3==1.24.1
websocket-client==0.55.0
